Abstract out fork and exec pattern
Abstracting out the fork and exec pattern allow different backends for
each, making the code more extensible.
